backend/requirements.txt
fastapi
uvicorn[standard]
paramiko
python-multipart
jinja2
pydantic-settings
passlib
pyqrcode
pypng